refactor: Move version package from internal to public module
The desktop module (a separate Go module) cannot import
internal packages from the backend module. Move the version
package from internal/util/version to util/version so it is
accessible to both modules.
