+++ backend/internal/worker/service/sysinfo.go
@@ -5,7 +5,7 @@ import (
 	"runtime"
 
 	leapmuxv1 "github.com/leapmux/leapmux/generated/proto/leapmux/v1"
-	"github.com/leapmux/leapmux/internal/util/version"
+	"github.com/leapmux/leapmux/util/version"
 	"github.com/leapmux/leapmux/internal/worker/channel"
 )
